Split factory status observability seams
